+++ src/main/java/by/epamtc/digapply/dao/AbstractDao.java
@@ -1,7 +1,7 @@
 package by.epamtc.digapply.dao;
 
 import by.epamtc.digapply.entity.Identifiable;
-import by.epamtc.digapply.mapper.RowMapper;
+import by.epamtc.digapply.dao.mapper.RowMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
