+++ src/main/java/by/epamtc/digapply/dao/mapper/RowMapper.java
@@ -1,4 +1,4 @@
-package by.epamtc.digapply.mapper;
+package by.epamtc.digapply.dao.mapper;
 
 import by.epamtc.digapply.entity.Identifiable;
